Create workspace structure

+++ src/main/java/com/example/back/backend/infrastructure/jpa/DaoAdminJpa.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface DaoAdminJpa extends JpaRepository<Admin,Long> {
+public interface DaoAdminJpa extends JpaRepository<Admin,Integer> {
 
     @Query(value = """
             SELECT * FROM admin WHERE EMAIL = :email
